add agent model and delete no longer needed model/agent combined pg_grid.py
